rename package entities
