+++ RR-Module/src/main/java/by/nc/school/dev/dao/IDao.java
@@ -1,6 +1,6 @@
 package by.nc.school.dev.dao;
 
-import by.nc.school.dev.enitities.BaseEntity;
+import by.nc.school.dev.entities.BaseEntity;
 
 interface IDao<T extends BaseEntity> {
