+++ RR-Module/src/main/java/by/nc/school/dev/entities/Curator.java
@@ -1,4 +1,4 @@
-package by.nc.school.dev.enitities;
+package by.nc.school.dev.entities;
 
 public class Curator extends Tutor {
