+++ RR-Module/src/main/java/by/nc/school/dev/entities/Dean.java
@@ -1,14 +1,14 @@
-package by.nc.school.dev.enitities;
+package by.nc.school.dev.entities;
 
-public class Tutor extends User {
+public class Dean extends User {
 
-    public Tutor(int id, String fullName, int role) {
+    public Dean(int id, String fullName, int role) {
         super(id, fullName, role);
     }
 
     @Override
     public String toString() {
-        return "Tutor{" +
+        return "Dean{" +
                 "fullName='" + fullName + '\'' +
                 ", role=" + role +
                 '}';
