+++ RR-Module/src/main/java/by/nc/school/dev/entities/Subject.java
@@ -1,4 +1,4 @@
-package by.nc.school.dev.enitities;
+package by.nc.school.dev.entities;
 
 public class Subject extends BaseEntity {
