+++ RR-Module/src/main/java/by/nc/school/dev/entities/Tutor.java
@@ -1,14 +1,14 @@
-package by.nc.school.dev.enitities;
+package by.nc.school.dev.entities;
 
-public class Dean extends User {
+public class Tutor extends User {
 
-    public Dean(int id, String fullName, int role) {
+    public Tutor(int id, String fullName, int role) {
         super(id, fullName, role);
     }
 
     @Override
     public String toString() {
-        return "Dean{" +
+        return "Tutor{" +
                 "fullName='" + fullName + '\'' +
                 ", role=" + role +
                 '}';
