+++ RR-Module/src/main/java/by/nc/school/dev/entities/User.java
@@ -1,4 +1,4 @@
-package by.nc.school.dev.enitities;
+package by.nc.school.dev.entities;
 
 public class User extends BaseEntity {
