+++ RR-Module/src/main/java/by/nc/school/dev/entities/WorkPlan.java
@@ -1,4 +1,4 @@
-package by.nc.school.dev.enitities;
+package by.nc.school.dev.entities;
 
 import by.nc.school.dev.utils.Pair;
