+++ RR-Module/src/main/java/by/nc/school/dev/entities/factories/CuratorFactory.java
@@ -1,7 +1,7 @@
-package by.nc.school.dev.enitities.factories;
+package by.nc.school.dev.entities.factories;
 
-import by.nc.school.dev.enitities.Curator;
-import by.nc.school.dev.enitities.User;
+import by.nc.school.dev.entities.Curator;
+import by.nc.school.dev.entities.User;
 
 public class CuratorFactory extends AbstractUserFactory {
