+++ RR-Module/src/main/java/by/nc/school/dev/entities/factories/DeanFactory.java
@@ -1,7 +1,7 @@
-package by.nc.school.dev.enitities.factories;
+package by.nc.school.dev.entities.factories;
 
-import by.nc.school.dev.enitities.Dean;
-import by.nc.school.dev.enitities.User;
+import by.nc.school.dev.entities.Dean;
+import by.nc.school.dev.entities.User;
 
 public class DeanFactory extends AbstractUserFactory {
