+++ RR-Module/src/main/java/by/nc/school/dev/entities/factories/UserFactory.java
@@ -1,7 +1,7 @@
-package by.nc.school.dev.enitities.factories;
+package by.nc.school.dev.entities.factories;
 
 import by.nc.school.dev.Role;
-import by.nc.school.dev.enitities.User;
+import by.nc.school.dev.entities.User;
 
 import java.util.HashMap;
 import java.util.Map;
